Excel

+++ Aban360.ReportPool.Persistence/Features/BuiltIns/PaymentTransactions/Implementations/WaterPaymentDetailQueryService.cs
@@ -34,7 +34,7 @@ namespace Aban360.ReportPool.Persistence.Features.BuiltIns.PaymentTransactions.I
                 ToAmount = input.ToAmount,
                 ReportDateJalali=DateTime.Now.ToShortPersianDateString(),
                 RecordCount = (waterPaymentDetailData is not null && waterPaymentDetailData.Any()) ? waterPaymentDetailData.Count() : 0,
-                TotalAmount = waterPaymentDetailData.Sum(payment => Convert.ToInt32(payment.Amount)),
+                TotalAmount = waterPaymentDetailData.Sum(payment => Convert.ToInt64(payment.Amount)),
             };
 
             var result = new ReportOutput<PaymentDetailHeaderOutputDto, PaymentDetailDataOutputDto>(ReportLiterals.WaterPaymentDetail, waterPaymentDetailHeader, waterPaymentDetailData);
